Refactor helper to create safe gufunc signature
